mcp_副本2/requirements.txt
fastmcp>=0.4.0
httpx>=0.27.0
python-dotenv>=1.0.1